This cl includes, removing the errors for sharing the files over distributive computing

+++ headers/clientsCommand.hpp
@@ -39,8 +39,8 @@ struct distributiveComputingargs
 {
     char *codePath;
     char *trainFilePath;
-    off_t train_offset;
-    size_t train_chunk_size;
+    long int train_offset;
+    long int train_chunk_size;
     char *IP;
 };
